Make the 256-bit P-256 a “runtime” alternative.
Fast implementation of primitives are needed in some cases, but we also
worry about code size on mobile devices. This change lays the groundwork
for alternative implementations: implementations that can be installed
at runtime. If the call to install isn't made, the linker should be able
to discard the unused code.

Change-Id: I264f04a0e1a9f77249573bcfb185e0da86522073