Update patch set 2
Patch Set 2:

(1 comment)

Hrm. Are you sure a build flag wouldn't be less hassle? This would suddenly do weird things if, say, we ever need to ship as a (private, of course) shared library.

How does the system OpenSSL on Android work? It's not exposed as part of the public API, but I think it's still a .so file.

I also keep getting pestered about the possibility of sharing a copy of BoringSSL in things which distribute together a copy of Cronet and another copy OpenSSL for other reasons, though whether that would look like a static linking or a private shared library, I'm not sure. I don't think it's actually happened for anything yet.

Patch-set: 2
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0