Update patch set 1
Patch Set 1:

In order to understand this change, it's good to look at the 11 other changes in this topic branch that follow it. In total, the first 11 changesets in this topic branch removes ~800 lines of non-functional code from crypto/ec, and the 12th changeset removes ~200 more assuming some of the unused functions exported in openssl/ec.h can be removed.

Patch-set: 1
Label: Code-Review=0