Update patch set 4
Patch Set 4:

> That function could use more love, but maybe it's later in this
 > series :)

Not in this series, though this function is quite crazy. I mostly did this because I was trying to document the callback and had a really rough time with the bounds. I'll probably cycle back to it when handshake messages get CBB'd.

Patch-set: 4
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0