Require source files define __STDC_FORMAT_MACROS to use BN FMT macros.
inttypes.h kindly requires a feature macro on some platforms. As bn.h is
a public header, we must leak this wart to the consumer. On platforms with
unfriendly inttypes.h headers, using BN_DEC_FMT1 and friends now require
the feature macro be defined externally.

This broke the Chromium Android Clang builder:
http://build.chromium.org/p/chromium.linux/builders/Android%20Clang%20Builder%20%28dbg%29/builds/59288

Change-Id: I88275a6788c7babd0eae32cae86f115bfa93a591