Don't go past the end of the buffer in generic RC4 implementation.
The generic RC4 implementation may read and write just past the end of the
buffer. It appropriately masks off and preserves the bytes off the end, so this
can only have practical effects if it crosses a page boundary. There's an
alignment check, so that can't happen; page boundaries are always aligned. But
it makes ASan unhappy and strictly speaking is a memory error.

Instead, fall through to the generic codepath which just reads it byte by byte.
This should fix the other bot failure.

Change-Id: I3cbd3bfc6cb0537e87f3252dea12d40ffa78d590