Fix SSL_get_current_cipher.
SSL_get_current_cipher is documented by upstream to return the cipher actually
being used. It currently doesn't return anything reasonable before the
handshake since it reads from the session.

Logic around s->session needs more comprehensive cleanup but, in the meantime,
defining it to be the current outgoing cipher is close to the current semantics
but for fixing the initial state (s->session->cipher is populated when sending
CCS). Store it in the SSL_AEAD_CTX which seems a natural place to associate
state pertaining to a connection half.

BUG=484744

Change-Id: Ife8db27a16615d0dbb2aec65359537243e08af7c