Reject overlarge scalars in ECC.
Brian Smith points out that the ECDSA code was consistently using
scalars that were greater than the group size for no good reason. This
change gets rid of that and also makes using overlarge scalars an error.

The test for “overlarge” is that they don't fit in the expected number
of limbs, not that they are greater than or equal to the order. So you
can use a scalar of p+1 because the important thing is the size.

This error behaviour is reflected in the general wNAF code too; not
because it's constant time but so that the behaviour be consistent
across platforms.

BUG=481190

Change-Id: Iea3f6403958af8d9b4cdaf04ffe6438b0b8206c2