Update patch set 3
Patch Set 3: Code-Review+2

> I know that's not the software you're aiming at, but nginx uses
> this.
> 
> There are obvious performance gains from avoiding extra read()
> syscalls.

nginx isn't going break with this change although performance isn't something to shed lightly.

None the less, this buffer code is such a mess that major work needs doing and reducing the number of modes is important.

We too need to do something about the overhead of the OpenSSL buffering model for our servers, but I think the answer for high performance situations is a very different API, which we should be able to add cleanly in the future, that avoids the copies altogether. (And, indeed, gets the ssl/ code out of the business of worrying about networking.)

Patch-set: 3
Label: Code-Review=+2