Handle endianness better in base.h.
For bi-endian architectures, use the endian-specific CPU feature flags.
(Also remove a duplicated aarch64 block.) Each of these are used in one
or both of Chromium and V8's build_config.h:
https://code.google.com/p/chromium/codesearch#chromium/src/v8/src/base/build_config.h&l=39
https://code.google.com/p/chromium/codesearch#chromium/src/build/build_config.h&sq=package:chromium&q=build/build_config.h&type=cs&l=111

Also NSPR, but we don't build that on ARM in Chromium:
https://code.google.com/p/chromium/codesearch#chromium/usr/include/nspr/prcpucfg.h&l=501
https://code.google.com/p/chromium/codesearch#chromium/usr/include/nspr/prcpucfg.h&l=553
https://code.google.com/p/chromium/codesearch#chromium/usr/include/nspr/prcpucfg.h&l=605

All supported targets are little-endian, but go ahead and condition
p256-64.c on little-endian so that doesn't silently break if it somehow
changes.

Change-Id: Ibcbc9869b70e617b4f904192d2d14942f8715788