Updates to the build-file generation script.
This change adds ‘gyp-standalone’ and ‘bazel’ targets to the build-file
generation script. It also allows multiple build outputs to be written
in one invocation, which saves the cost of Perl to build all the
assembly files multiple times.

Note that this will require changes to the Chromium GYP files when its
rolled out because “boringssl_lib_sources” has been broken into
“boringssl_crypto_sources” and “boringssl_ssl_sources”.

Change-Id: Icfe9678d08fba74fd215cf92521831e9cdc265da