Update patch set 1
Patch Set 1:

New post, as I can't find how to edit the previous. Description here, as the upload UI didn't allow me to add one.

Rationale: The bit strength of the handshake is of importance to the connection security, as shown e.g. by logjam. In particular for logjam, 1024 bit DH is far from optimal security. Make this information available to API customers, to allow them to handle this.

Possible actions taken on the new information are:
* Statistics about how often weak parameters (read: DH 1024) is used
* Indicate this in advanced UI, along with other connection parameters
* Make security decisions, such as silently downgrading the UI from max security to dubios/web.

Patch-set: 1
Label: Code-Review=0