Update patch set 3
Patch Set 3:

I am planning to surface this in the UI in Opera. I can't speak for Chrome. Opera currently has a string like "TLS 1.2 AES_128_CBS SHA1 DHE_RSA". The plan is to add the key exchange, e.g. like "DHE_1024_RSA". Concerned users will then at least have a method for telling the security of a connection (while to the vast majority of users it continues to be a unintelligble string of gobbledygook). This is a fix wanted before logjam (Opera 12 shows this information), but which made it to the top of the priority list following logjam. It is also to be expected that some of the other, currently reliable key exchanges will turn weak in some time, and it is of value to webmasters to get this information.

We are thinking of logging key exchange strength in statistics (in particular the use of DH 1024, possibly combined with EV), and also of softly downgrading the security of DHE 1024. (If Chromium blocks DHE entirely, then no need for this currently, but as far as I understand that decision hasn't been made yet.) For instance start by downgrading EV DH 1024 to DV UI, and make a long term plan for gradually phasing out DH 1024 (step by step, presumably many steps including UI only).

It would presumably also make sense to make this available in Developer Tools, but that would be a pure Chromium patch which I do not intend to write.

Patch-set: 3