Zero memory in |OPENSSL_free|.
This change causes |OPENSSL_free| to zero memory before |free| is
called. This will systematically clean memory, as opposed to trying to
make sure that |OPENSSL_cleanse| is called at every place needed, which
is error prone.

The downside is that if users get a pointer from BoringSSL and call
|free| rather than |OPENSSL_free| on it, they're going to have a bad
time.

Change-Id: Ibc4533934f373696311e6f57298c2da57dfbd772