Update patch set 1
Patch Set 1:

(13 comments)

> This is a bit of an exploratory change. See what you think.

Hrm. If we don't want to give up on _cleanse altogether, this seems a good way to do it. If not this, being cleanse-clean will at least requiring pushing it down into CBB (I'm thinking the SSL_SESSION serialization case).

The one thing that might be sad is SSL stack's large-ish record buffers. I don't think those ever contain data we ever care about cleansing (or, if we do, we can only cleanse as far as we wrote to) and we'll bounce them in and out a lot for even small reads and writes. Do you think that matters? We could add OPENSSL_free_uncleansed or something for those if it does.

I think everything else should be reasonably amortizable with the cost of filling in the data. Those buffers are the only place I can think of where we routinely allocate a much larger buffer than we may actually need.

(Also might be something to be said for pushing the buffer code up. I should get on that buffer-free logic. Been sitting on a half-finished branch that gets it out of ssl3_get_record.)

Patch-set: 1