Clarify control flow in |bn_mul_mont| in bn/generic.c.
When building in OPENSSL_NO_ASM mode, MSVC complains about unreachable
code. The redundant initialization of |i| is the main problem. The skipping of
the first test of the condition |i < num| with |goto| was also confusing. This
change replaces the |goto| with a normal |if| condition inside the loop with
the expectation that the compiler will likely generate the same code as before.

Change-Id: Id551899b2602824978edc1a1cb0703b76516808d